requirements.txt
flask
tensorflow-cpu
opencv-python
numpy
flask-socketio
pillow
imutils
flask_cors